+++ backend/src/main/java/sullog/backend/member/config/jwt/JwtAuthFilter.java
@@ -5,7 +5,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
-import sullog.backend.member.service.TokenService;
+import sullog.backend.auth.service.TokenService;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletException;
